[UPDATE] the sql bind to work in threadsafe mode
Update the totality of the sql command
Did a complete overhall of the sql bind
Need to implement tests for the sql library
